move all loading/writing functions to a central place
