+++ bin/BarTracker
@@ -12,6 +12,7 @@ from madmom.audio.signal import SignalProcessor
 from madmom.features import ActivationsProcessor
 from madmom.features.downbeats import (DBNBarTrackingProcessor,
                                        LoadBeatsProcessor, RNNBarProcessor)
+from madmom.io import write_beats
 from madmom.processors import IOProcessor, ParallelProcessor, io_arguments
 
 
@@ -100,19 +101,13 @@ def main():
         # save the RNN bar activations to file
         out_processor = ActivationsProcessor(mode='w', **vars(args))
     else:
-        # detect the downbeats in the bar activation function
+        # track the (down-)beats in the bar activation function and output them
         out_processor = [DBNBarTrackingProcessor(**vars(args))]
-        # append writer
         if args.downbeats:
             # filter the beats to contain only downbeats
             from madmom.features.downbeats import filter_downbeats
             out_processor.append(filter_downbeats)
-            # simply write the timestamps
-            from madmom.utils import write_events as writer
-        else:
-            # borrow the note writer for outputting timestamps + beat numbers
-            from madmom.features.notes import write_notes as writer
-        out_processor.append(writer)
+        out_processor.append(write_beats)
 
     # create an IOProcessor
     processor = IOProcessor(in_processor, out_processor)
