+++ bin/CRFBeatDetector
@@ -13,6 +13,7 @@ from madmom.audio.signal import SignalProcessor
 from madmom.features import ActivationsProcessor
 from madmom.features.beats import RNNBeatProcessor, CRFBeatDetectionProcessor
 from madmom.features.tempo import TempoEstimationProcessor
+from madmom.io import write_beats
 
 
 def main():
@@ -94,12 +95,9 @@ def main():
         # save the RNN beat activations to file
         out_processor = ActivationsProcessor(mode='w', **vars(args))
     else:
-        # detect the beats with a CRF
+        # detect the beats with a CRF and output them
         beat_processor = CRFBeatDetectionProcessor(**vars(args))
-        # output handler
-        from madmom.utils import write_events as writer
-        # sequentially process them
-        out_processor = [beat_processor, writer]
+        out_processor = [beat_processor, write_beats]
 
     # create an IOProcessor
     processor = IOProcessor(in_processor, out_processor)
