+++ bin/DBNDownBeatTracker
@@ -14,6 +14,7 @@ from madmom.audio.signal import SignalProcessor
 from madmom.features import ActivationsProcessor
 from madmom.features.downbeats import (RNNDownBeatProcessor,
                                        DBNDownBeatTrackingProcessor)
+from madmom.io import write_beats
 
 
 def main():
@@ -81,19 +82,13 @@ def main():
         # save the RNN beat activations to file
         out_processor = ActivationsProcessor(mode='w', **vars(args))
     else:
-        # track the beats with a DBN
+        # track the (down-)beats with a DBN and output them
         out_processor = [DBNDownBeatTrackingProcessor(**vars(args))]
         if args.downbeats:
             # filter the beats to contain only downbeats
             from madmom.features.downbeats import filter_downbeats
             out_processor.append(filter_downbeats)
-            # simply write the timestamps
-            from madmom.utils import write_events as writer
-        else:
-            # borrow the note writer for outputting timestamps + beat numbers
-            from madmom.features.notes import write_notes as writer
-        # sequentially process them
-        out_processor.append(writer)
+        out_processor.append(write_beats)
 
     # create an IOProcessor
     processor = IOProcessor(in_processor, out_processor)
