+++ bin/MMBeatTracker
@@ -14,6 +14,7 @@ from madmom.audio.signal import SignalProcessor
 from madmom.features import ActivationsProcessor
 from madmom.features.beats import (RNNBeatProcessor, DBNBeatTrackingProcessor,
                                    MultiModelSelectionProcessor)
+from madmom.io import write_beats
 
 
 def main():
@@ -91,12 +92,9 @@ def main():
         # save the RNN beat activations to file
         out_processor = ActivationsProcessor(mode='w', **vars(args))
     else:
-        # track the beats with a DBN
+        # track the beats with a DBN and output them
         beat_processor = DBNBeatTrackingProcessor(**vars(args))
-        # output handler
-        from madmom.utils import write_events as writer
-        # sequentially process them
-        out_processor = [beat_processor, writer]
+        out_processor = [beat_processor, write_beats]
 
     # create an IOProcessor
     processor = IOProcessor(in_processor, out_processor)
