+++ bin/PianoTranscriptor
@@ -13,8 +13,8 @@ from madmom.processors import IOProcessor, io_arguments
 from madmom.audio.signal import SignalProcessor
 from madmom.features import ActivationsProcessor
 from madmom.features.notes import (RNNPianoNoteProcessor,
-                                   NotePeakPickingProcessor, write_midi,
-                                   write_notes, write_mirex_format)
+                                   NotePeakPickingProcessor)
+from madmom.io import write_notes, write_midi, write_mirex_format
 
 
 def main():
